Break up super.c into a few modules
