+++ src/super.c
@@ -4,8 +4,8 @@
  * @date Created  on Aug 24, 2024
  * @author Attila Kovacs
  *
- *  SuperNOVAS only functions, which are not integral to the functionality of novas.c, and thus
- *  can live in a separate, more manageably sized, module.
+ *  Various SuperNOVAS only functions, which are not integral to the functionality of novas.c,
+ *  and thus can live in a separate, more manageably sized, module.
  */
 
 // We'll use gcc major version as a proxy for the glibc library to decide which feature macro to use.
@@ -33,836 +33,7 @@ int strcasecmp(const char *s1, const char *s2);
 
 /// \endcond
 
-/**
- * Returns the difference between Terrestrial Time (TT) and Universal Coordinated Time (UTC)
- *
- * @param leap_seconds  [s] The current leap seconds (see IERS Bulletins)
- * @return              [s] The TT - UTC time difference
- *
- * @sa get_ut1_to_tt()
- * @sa julian_date()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-double get_utc_to_tt(int leap_seconds) {
-  return leap_seconds + NOVAS_TAI_TO_TT;
-}
 
-/**
- * Returns the TT - UT1 time difference given the leap seconds and the actual UT1 - UTC time
- * difference as measured and published by IERS.
- *
- * NOTES:
- * <ol>
- * <li>The current UT1 - UTC time difference, and polar offsets, historical data and near-term
- * projections are published in the
- <a href="https://www.iers.org/IERS/EN/Publications/Bulletins/bulletins.html>IERS Bulletins</a>
- * </li>
- * </ol>
- *
- * @param leap_seconds  [s] Leap seconds at the time of observations
- * @param dut1          [s] UT1 - UTC time difference [-0.5:0.5]
- * @return              [s] The TT - UT1 time difference that is suitable for used with all
- *                      calls in this library that require a <code>ut1_to_tt</code> argument.
- *
- * @sa get_utc_to_tt()
- * @sa place()
- * @sa cel_pole()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-double get_ut1_to_tt(int leap_seconds, double dut1) {
-  return get_utc_to_tt(leap_seconds) + dut1;
-}
-
-/**
- * Computes the International Celestial Reference System (ICRS) position of a source.
- * (from the geocenter). Unlike `place_gcrs()`, this version does not include
- * aberration or gravitational deflection corrections.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date of observation.
- * @param source      Catalog source or solar_system body.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param[out] pos    Structure to populate with the calculated geocentric ICRS position
- *                    data (Unlike place_gcrs(), the calculated coordinates do not account
- *                    for aberration or gravitational deflection).
- * @return            0 if successful, or -1 if any of the input pointer arguments is NULL,
- *                    or else an error from place().
- *
- * @sa place_gcrs()
- * @sa place_cirs()
- * @sa place_tod()
- * @sa mean_star()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int place_icrs(double jd_tt, const object *restrict source, enum novas_accuracy accuracy, sky_pos *restrict pos) {
-  prop_error("place_icrs", place(jd_tt, source, NULL, 0.0, NOVAS_ICRS, accuracy, pos), 0);
-  return 0;
-}
-
-/**
- * Computes the Geocentric Celestial Reference System (GCRS) position of a source (as 'seen'
- * from the geocenter) at the given time of observation. Unlike `place_icrs()`, this includes
- * aberration for the moving frame of the geocenter as well as gravitational deflections
- * calculated for a virtual observer located at the geocenter. See `place()` for more information.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date of observation.
- * @param source      Catalog source or solar_system body.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param[out] pos    Structure to populate with the calculated GCRS position data
- * @return            0 if successful, or -1 if any of the input pointer arguments is NULL,
- *                    or else an error from place().
- *
- * @sa place_icrs()
- * @sa place_cirs()
- * @sa place_tod()
- * @sa virtual_star()
- * @sa virtual_planet()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int place_gcrs(double jd_tt, const object *restrict source, enum novas_accuracy accuracy, sky_pos *restrict pos) {
-  prop_error("place_gcrs", place(jd_tt, source, NULL, 0.0, NOVAS_GCRS, accuracy, pos), 0);
-  return 0;
-}
-
-/**
- * Computes the Celestial Intermediate Reference System (CIRS) dynamical position
- * position of a source as 'seen' from the geocenter at the given time of observation. See
- * `place()` for more information.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date of observation.
- * @param source      Catalog source or solar_system body.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param[out] pos    Structure to populate with the calculated CIRS position data
- * @return            0 if successful, or -1 if any of the input pointer arguments is NULL,
- *                    or else an error from place().
- *
- * @sa place_tod()
- * @sa place_gcrs()
- *
- * @since 1.0
- * @author Attila Kovacs
- *
- */
-int place_cirs(double jd_tt, const object *restrict source, enum novas_accuracy accuracy, sky_pos *restrict pos) {
-  prop_error("place_cirs", place(jd_tt, source, NULL, 0.0, NOVAS_CIRS, accuracy, pos), 0);
-  return 0;
-}
-
-/**
- * Computes the True of Date (TOD) dynamical position position of a source as 'seen' from the
- * geocenter at the given time of observation. See `place()` for more information.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date of observation.
- * @param source      Catalog source or solar_system body.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param[out] pos    Structure to populate with the calculated CIRS position data
- * @return            0 if successful, or -1 if any of the input pointer arguments is NULL,
- *                    or else an error from place().
- *
- * @sa place_cirs()
- * @sa place_gcrs()
- * @sa app_star()
- * @sa app_planet()
- *
- * @since 1.0
- * @author Attila Kovacs
- *
- */
-int place_tod(double jd_tt, const object *restrict source, enum novas_accuracy accuracy, sky_pos *restrict pos) {
-  prop_error("place_tod", place(jd_tt, source, NULL, 0.0, NOVAS_TOD, accuracy, pos), 0);
-  return 0;
-}
-
-/**
- * Computes the Mean of Date (MOD) dynamical position position of a source as 'seen' from the
- * geocenter at the given time of observation. See `place()` for more information.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date of observation.
- * @param source      Catalog source or solar_system body.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param[out] pos    Structure to populate with the calculated CIRS position data
- * @return            0 if successful, or -1 if any of the input pointer arguments is NULL,
- *                    or else an error from place().
- *
- * @sa place_cirs()
- * @sa place_gcrs()
- * @sa app_star()
- * @sa app_planet()
- *
- * @since 1.1
- * @author Attila Kovacs
- *
- */
-int place_mod(double jd_tt, const object *restrict source, enum novas_accuracy accuracy, sky_pos *restrict pos) {
-  prop_error("place_mod", place(jd_tt, source, NULL, 0.0, NOVAS_MOD, accuracy, pos), 0);
-  return 0;
-}
-
-/**
- * Computes the J2000 dynamical position position of a source as 'seen' from the
- * geocenter at the given time of observation. See `place()` for more information.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date of observation.
- * @param source      Catalog source or solar_system body.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param[out] pos    Structure to populate with the calculated CIRS position data
- * @return            0 if successful, or -1 if any of the input pointer arguments is NULL,
- *                    or else an error from place().
- *
- * @sa place_cirs()
- * @sa place_gcrs()
- * @sa app_star()
- * @sa app_planet()
- *
- * @since 1.1
- * @author Attila Kovacs
- *
- */
-int place_j2000(double jd_tt, const object *restrict source, enum novas_accuracy accuracy, sky_pos *restrict pos) {
-  prop_error("place_j2000", place(jd_tt, source, NULL, 0.0, NOVAS_J2000, accuracy, pos), 0);
-  return 0;
-}
-
-/**
- * Convert ecliptic longitude and latitude to right ascension and declination.  To convert
- * GCRS ecliptic coordinates (mean ecliptic and equinox of J2000.0), set 'coord_sys' to
- * NOVAS_GCRS_EQUATOR(2); in this case the value of 'jd_tt' can be set to anything, since
- * J2000.0 is assumed. Otherwise, all input coordinates are dynamical at'jd_tt'.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date. (Unused if 'coord_sys'
- *                    is NOVAS_GCRS_EQUATOR[2])
- * @param coord_sys   The astrometric reference system of the coordinates. If 'coord_sys' is
- *                    NOVAS_GCRS_EQUATOR(2), the input GCRS coordinates are converted to
- *                    J2000 ecliptic coordinates.
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param elon        [deg] Ecliptic longitude in degrees, referred to specified ecliptic and
- *                    equinox of date.
- * @param elat        [deg] Ecliptic latitude in degrees, referred to specified ecliptic and
- *                    equinox of date.
- * @param[out] ra     [h] Right ascension in hours, referred to specified equator and equinox
- *                    of date.
- * @param[out] dec    [deg] Declination in degrees, referred to specified equator and equinox
- *                    of date.
-
- * @return            0 if successful, or else 1 if the value of 'coord_sys' is invalid.
- *
- * @sa ecl2equ_vec()
- * @sa equ2ecl()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int ecl2equ(double jd_tt, enum novas_equator_type coord_sys, enum novas_accuracy accuracy, double elon, double elat,
-        double *restrict ra, double *restrict dec) {
-  static const char *fn = "ecl2equ";
-  double coslat, pos[3], xyproj;
-
-  if(!ra || !dec)
-    return novas_error(-1, EINVAL, fn, "NULL output pointer: ra=%p, dec=%p", ra, dec);
-
-  // Form position vector in equatorial system from input coordinates.
-  elon *= DEGREE;
-  elat *= DEGREE;
-  coslat = cos(elat);
-
-  pos[0] = coslat * cos(elon);
-  pos[1] = coslat * sin(elon);
-  pos[2] = sin(elat);
-
-  // Convert the vector from equatorial to ecliptic system.
-  prop_error(fn, ecl2equ_vec(jd_tt, coord_sys, accuracy, pos, pos), 0);
-
-  // Decompose ecliptic vector into ecliptic longitude and latitude.
-  xyproj = sqrt(pos[0] * pos[0] + pos[1] * pos[1]);
-  *ra = xyproj ? atan2(pos[1], pos[0]) / HOURANGLE : 0.0;
-  if(*ra < 0.0)
-    *ra += DAY_HOURS;
-
-  *dec = atan2(pos[2], xyproj) / DEGREE;
-
-  return 0;
-}
-
-/**
- * Converts galactic longitude and latitude to ICRS right ascension and declination.
- *
- * REFERENCES:
- * <ol>
- * <li>Hipparcos and Tycho Catalogues, Vol. 1, Section 1.5.3.</li>
- * </ol>
- *
- * @param glon        [deg] Galactic longitude in degrees.
- * @param glat        [deg] Galactic latitude in degrees.
- * @param[out] ra     [h] ICRS right ascension in hours.
- * @param[out] dec    [deg] ICRS declination in degrees.
-
- *
- * @return            0 if successful, or -1 if either of the output pointer arguments
- *                    are NULL.
- *
- * @sa equ2gal()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int gal2equ(double glon, double glat, double *restrict ra, double *restrict dec) {
-  double pos1[3], pos2[3], xyproj, coslat;
-
-  // Rotation matrix A_g from Hipparcos documentation eq. 1.5.11.
-  // AK: Transposed compared to NOVAS C 3.1 for dot product handling.
-  static const double ag[3][3] = { //
-          { -0.0548755604, +0.4941094279, -0.8676661490 }, //
-          { -0.8734370902, -0.4448296300, -0.1980763734 }, //
-          { -0.4838350155, +0.7469822445, +0.4559837762 } };
-
-  if(!ra || !dec)
-    return novas_error(-1, EINVAL, "gal2equ", "NULL output pointer: ra=%p, dec=%p", ra, dec);
-
-  // Form position vector in equatorial system from input coordinates
-  glon *= DEGREE;
-  glat *= DEGREE;
-  coslat = cos(glat);
-
-  pos1[0] = coslat * cos(glon);
-  pos1[1] = coslat * sin(glon);
-  pos1[2] = sin(glat);
-
-  // Rotate position vector to galactic system, using Hipparcos documentation eq. 1.5.13.
-  pos2[0] = novas_vdot(ag[0], pos1);
-  pos2[1] = novas_vdot(ag[1], pos1);
-  pos2[2] = novas_vdot(ag[2], pos1);
-
-  // Decompose galactic vector into longitude and latitude.
-  xyproj = sqrt(pos2[0] * pos2[0] + pos2[1] * pos2[1]);
-  *ra = xyproj ? atan2(pos2[1], pos2[0]) / HOURANGLE : 0.0;
-  if(*ra < 0.0)
-    *ra += DAY_HOURS;
-
-  *dec = atan2(pos2[2], xyproj) / DEGREE;
-
-  return 0;
-}
-
-/**
- * Change J2000 coordinates to GCRS coordinates. Same as frame_tie() called with J2000_TO_ICRS
- *
- * @param in        J2000 input 3-vector
- * @param[out] out  GCRS output 3-vector
- * @return          0 if successful, or else an error from frame_tie()
- *
- * @sa j2000_to_tod()
- * @sa gcrs_to_j2000()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int j2000_to_gcrs(const double *in, double *out) {
-  prop_error("j2000_to_gcrs", frame_tie(in, J2000_TO_ICRS, out), 0);
-  return 0;
-}
-
-/**
- * Transforms a rectangular equatorial (x, y, z) vector from the Celestial Intermediate
- * Reference System (CIRS) at the given epoch to the True of Date (TOD) reference
- * system.
- *
- * @param jd_tt     [day] Terrestrial Time (TT) based Julian date that defines
- *                  the output epoch. Typically it does not require much precision, and
- *                  Julian dates in other time measures will be unlikely to affect the
- *                  result
- * @param accuracy  NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param in        CIRS Input (x, y, z) position or velocity vector
- * @param[out] out  Output position or velocity 3-vector in the True of Date (TOD) frame.
- *                  It can be the same vector as the input.
- * @return          0 if successful, or -1 if either of the vector arguments is NULL
- *                  or the accuracy is invalid, or 10 + the error from cio_location(), or
- *                  else 20 + the error from cio_basis().
- *
- * @sa tod_to_cirs()
- * @sa cirs_to_app_ra()
- * @sa cirs_to_gcrs()
- * @sa cirs_to_itrs()
- *
- *
- * @since 1.1
- * @author Attila Kovacs
- */
-int cirs_to_tod(double jd_tt, enum novas_accuracy accuracy, const double *in, double *out) {
-  static const char *fn = "cirs_to_tod";
-
-  double ra_cio;  // [h] R.A. of the CIO (from the true equinox) we'll calculate
-
-  // Obtain the R.A. [h] of the CIO at the given date
-  prop_error(fn, cio_ra(jd_tt, accuracy, &ra_cio), 0);
-  prop_error(fn, spin(-15.0 * ra_cio, in, out), 0);
-
-  return 0;
-}
-
-/**
- * Transforms a rectangular equatorial (x, y, z) vector from the True of Date (TOD) reference
- * system to the Celestial Intermediate Reference System (CIRS) at the given epoch to the .
- *
- * NOTES:
- * <ol>
- * <li>The accuracy of the output CIRS coordinates depends on how the input TOD coordinates
- * were obtained. If TOD was calculated via the old (pre IAU 2006) method, using the Lieske et
- * al. 1997 nutation model, then the limited accuracy of that model will affect the resulting
- * coordinates. This is the case for the SuperNOVAS functions novas_geom_posvel() and
- * novas_sky_pos() also, when called with `NOVAS_TOD` as the system, as well as all legacy NOVAS
- * C functions that produce TOD coordinates.</li>
- * </ol>
- *
- * @param jd_tt     [day] Terrestrial Time (TT) based Julian date that defines
- *                  the output epoch. Typically it does not require much precision, and
- *                  Julian dates in other time measures will be unlikely to affect the
- *                  result
- * @param accuracy  NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param in        CIRS Input (x, y, z) position or velocity vector
- * @param[out] out  Output position or velocity 3-vector in the True of Date (TOD) frame.
- *                  It can be the same vector as the input.
- * @return          0 if successful, or -1 if either of the vector arguments is NULL
- *                  or the accuracy is invalid, or 10 + the error from cio_location(), or
- *                  else 20 + the error from cio_basis().
- *
- * @sa cirs_to_tod()
- * @sa app_to_cirs_ra()
- * @sa tod_to_gcrs()
- * @sa tod_to_j2000()
- * @sa tod_to_itrs()
- *
- *
- * @since 1.1
- * @author Attila Kovacs
- */
-int tod_to_cirs(double jd_tt, enum novas_accuracy accuracy, const double *in, double *out) {
-  static const char *fn = "tod_to_cirs";
-
-  double ra_cio;  // [h] R.A. of the CIO (from the true equinox) we'll calculate
-
-  // Obtain the R.A. [h] of the CIO at the given date
-  prop_error(fn, cio_ra(jd_tt, accuracy, &ra_cio), 0);
-  prop_error(fn, spin(15.0 * ra_cio, in, out), 0);
-
-  return 0;
-}
-
-
-/**
- * Converts a position vector in the Earth-fixed ITRS frame to astrometric (unrefracted) azimuth
- * and zenith angles at the specified observer location.
- *
- * @param location   Observer location on Earth
- * @param itrs       3-vector position in Earth-fixed ITRS frame
- * @param[out] az    [deg] astrometric azimuth angle at observer location [0:360]. It may be
- *                   NULL if not required.
- * @param[out] za    [deg] astrometric zenith angle at observer location [0:180]. It may be NULL
- *                   if not required.
- * @return           0 if successful, or else -1 if the location or the input vector is NULL.
- *
- * @sa hor_to_itrs()
- * @sa cirs_to_itrs()
- * @sa tod_to_itrs()
- * @sa refract_astro()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int itrs_to_hor(const on_surface *restrict location, const double *restrict itrs, double *restrict az, double *restrict za) {
-  double uze[3], une[3], uwe[3];
-  double lat, lon, coslat, sinlat, coslon, sinlon;
-  double pn, pw, pz, proj;
-
-  // Default output values in case of error return.
-  if(az)
-    *az = NAN;
-  if(za)
-    *za = NAN;
-
-  if(!location || !itrs)
-    return novas_error(-1, EINVAL, "itrs_to_hor", "NULL input location=%p or ITRS pos=%p", location, itrs);
-
-  lat = location->latitude * DEGREE;
-  lon = location->longitude * DEGREE;
-
-  coslat = cos(lat);
-  sinlat = sin(lat);
-  coslon = cos(lon);
-  sinlon = sin(lon);
-
-  // Define vector toward local north in Earth-fixed system (x axis).
-  une[0] = -sinlat * coslon;
-  une[1] = -sinlat * sinlon;
-  une[2] = coslat;
-
-  // Define vector toward local west in Earth-fixed system (y axis).
-  uwe[0] = sinlon;
-  uwe[1] = -coslon;
-  uwe[2] = 0.0;
-
-  // Define vector toward local zenith in Earth-fixed system (z axis).
-  uze[0] = coslat * coslon;
-  uze[1] = coslat * sinlon;
-  uze[2] = sinlat;
-
-  // Obtain vectors in celestial system.
-
-  // Compute coordinates of object w.r.t orthonormal basis.
-
-  // Compute components of 'p' - projections of 'p' onto rotated
-  // Earth-fixed basis vectors.
-  pn = novas_vdot(itrs, une);
-  pw = novas_vdot(itrs, uwe);
-  pz = novas_vdot(itrs, uze);
-
-  // Compute azimuth and zenith distance.
-  proj = sqrt(pn * pn + pw * pw);
-
-  if(az) {
-    *az = proj > 0.0 ? -atan2(pw, pn) / DEGREE : 0.0;
-    if(*az < 0.0)
-      *az += DEG360;
-  }
-
-  if(za)
-    *za = atan2(proj, pz) / DEGREE;
-
-  return 0;
-}
-
-/**
- * Converts astrometric (unrefracted) azimuth and zenith angles at the specified observer location to a
- * unit position vector in the Earth-fixed ITRS frame.
- *
- * @param location   Observer location on Earth
- * @param az         [deg] astrometric azimuth angle at observer location [0:360]. It may be
- *                   NULL if not required.
- * @param za         [deg] astrometric zenith angle at observer location [0:180]. It may be NULL
- *                   if not required.
- * @param[out] itrs  Unit 3-vector direction in Earth-fixed ITRS frame
- * @return           0 if successful, or else -1 if the location or the input vector is NULL.
- *
- * @sa itrs_to_hor()
- * @sa itrs_to_cirs()
- * @sa itrs_to_tod()
- * @sa refract()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int hor_to_itrs(const on_surface *restrict location, double az, double za, double *restrict itrs) {
-  double in[3], uze[3], une[3], uwe[3];
-  double sinza;
-  double lat, lon, coslat, sinlat, coslon, sinlon;
-
-  if(!location || !itrs)
-    return novas_error(-1, EINVAL, "hor_to_itrs", "NULL input location=%p or output itrs=%p position", location, itrs);
-
-  az *= -DEGREE;
-  za *= DEGREE;
-
-  sinza = sin(za);
-  in[0] = sinza * cos(az);
-  in[1] = sinza * sin(az);
-  in[2] = cos(za);
-
-  lat = location->latitude * DEGREE;
-  lon = location->longitude * DEGREE;
-
-  coslat = cos(lat);
-  sinlat = sin(lat);
-  coslon = cos(lon);
-  sinlon = sin(lon);
-
-  // Define vector toward local zenith in Earth-fixed system (z axis).
-  uze[0] = coslat * coslon;
-  uze[1] = coslat * sinlon;
-  uze[2] = sinlat;
-
-  // Define vector toward local north in Earth-fixed system (x axis).
-  une[0] = -sinlat * coslon;
-  une[1] = -sinlat * sinlon;
-  une[2] = coslat;
-
-  // Define vector toward local west in Earth-fixed system (y axis).
-  uwe[0] = sinlon;
-  uwe[1] = -coslon;
-  uwe[2] = 0.0;
-
-  // Calculate ITRS positions from NWZ components at ITRS origin
-  itrs[0] = une[0] * in[0] + uwe[0] * in[1] + uze[0] * in[2];
-  itrs[1] = une[1] * in[0] + uwe[1] * in[1] + uze[1] * in[2];
-  itrs[2] = une[2] * in[0] + uwe[2] * in[1] + uze[2] * in[2];
-
-  return 0;
-}
-
-/**
- * Converts a CIRS right ascension coordinate (measured from the CIO) to an apparent R.A.
- * measured from the true equinox of date.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param ra          [h] The CIRS right ascension coordinate, measured from the CIO.
- * @return            [h] the apparent R.A. coordinate measured from the true equinox of date [0:24],
- *                    or NAN if the accuracy is invalid, or if there wan an error from cio_ra().
- *
- * @sa app_to_cirs_ra()
- * @sa cirs_to_tod()
- *
- * @since 1.0.1
- * @author Attila Kovacs
- */
-double cirs_to_app_ra(double jd_tt, enum novas_accuracy accuracy, double ra) {
-  double ra_cio;  // [h] R.A. of the CIO (from the true equinox) we'll calculate
-
-  // Obtain the R.A. [h] of the CIO at the given date
-  int stat = cio_ra(jd_tt, accuracy, &ra_cio);
-  if(stat)
-    return novas_trace_nan("cirs_to_app_ra");
-
-  // Convert CIRS R.A. to true apparent R.A., keeping the result in the [0:24] h range
-  ra = remainder(ra + ra_cio, DAY_HOURS);
-  if(ra < 0.0)
-    ra += DAY_HOURS;
-
-  return ra;
-}
-
-/**
- * Converts an apparent right ascension coordinate (measured from the true equinox of date) to a
- * CIRS R.A., measured from the CIO.
- *
- * @param jd_tt       [day] Terrestrial Time (TT) based Julian date
- * @param accuracy    NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param ra          [h] the apparent R.A. coordinate measured from the true equinox of date.
- * @return            [h] The CIRS right ascension coordinate, measured from the CIO [0:24],
- *                    or NAN if the accuracy is invalid, or if there wan an error from cio_ra().
- *
- * @sa cirs_to_app_ra()
- * @sa tod_to_cirs()
- *
- * @since 1.0.1
- * @author Attila Kovacs
- */
-double app_to_cirs_ra(double jd_tt, enum novas_accuracy accuracy, double ra) {
-  double ra_cio;  // [h] R.A. of the CIO (from the true equinox) we'll calculate
-
-  // Obtain the R.A. [h] of the CIO at the given date
-  int stat = cio_ra(jd_tt, accuracy, &ra_cio);
-  if(stat)
-    return novas_trace_nan("app_to_cirs_ra");
-
-  // Convert CIRS R.A. to true apparent R.A., keeping the result in the [0:24] h range
-  ra = remainder(ra - ra_cio, DAY_HOURS);
-  if(ra < 0.0)
-    ra += DAY_HOURS;
-
-  return ra;
-}
-
-/**
- * Rotates a position vector from the Earth-fixed ITRS frame to the dynamical CIRS frame of
- * date (IAU 2000 standard method).
- *
- * If both 'xp' and 'yp' are set to 0 no polar motion is included in the transformation.
- *
- * If extreme (sub-microarcsecond) accuracy is not required, you can use UT1-based Julian date
- * instead of the TT-based Julian date and set the 'ut1_to_tt' argument to 0.0. and you can
- * use UTC-based Julian date the same way.for arcsec-level precision also.
- *
- * REFERENCES:
- *  <ol>
- *   <li>Kaplan, G. H. et. al. (1989). Astron. Journ. 97, 1197-1210.</li>
- *   <li>Kaplan, G. H. (2003), 'Another Look at Non-Rotating Origins', Proceedings of IAU
- *   XXV Joint Discussion 16.</li>
- *  </ol>
- *
- * @param jd_tt_high    [day] High-order part of Terrestrial Time (TT) based Julian date.
- * @param jd_tt_low     [day] Low-order part of Terrestrial Time (TT) based Julian date.
- * @param ut1_to_tt     [s] TT - UT1 Time difference in seconds
- * @param accuracy      NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param xp            [arcsec] Conventionally-defined X coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param yp            [arcsec] Conventionally-defined Y coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param in            Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to ITRS axes (terrestrial system)
- * @param[out] out      Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to CIRS axes (celestial system).
- * @return              0 if successful, -1 if either of the vector arguments is NULL, 1 if
- *                      'accuracy' is invalid, or else 10 + the error from cio_location(), or
- *                      20 + error from cio_basis().
- *
- * @sa itrs_to_tod()
- * @sa cirs_to_itrs()
- * @sa cirs_to_gcrs()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int itrs_to_cirs(double jd_tt_high, double jd_tt_low, double ut1_to_tt, enum novas_accuracy accuracy, double xp, double yp,
-        const double *in, double *out) {
-  prop_error("itrs_to_cirs",
-          ter2cel(jd_tt_high, jd_tt_low - ut1_to_tt / DAY, ut1_to_tt, EROT_ERA, accuracy, NOVAS_DYNAMICAL_CLASS, xp, yp, in, out), 0);
-  return 0;
-}
-
-/**
- * Rotates a position vector from the Earth-fixed ITRS frame to the dynamical True of Date
- * (TOD) frame of date (pre IAU 2000 method).
- *
- * If both 'xp' and 'yp' are set to 0 no polar motion is included in the transformation.
- *
- * If extreme (sub-microarcsecond) accuracy is not required, you can use UT1-based Julian date
- * instead of the TT-based Julian date and set the 'ut1_to_tt' argument to 0.0. and you can
- * use UTC-based Julian date the same way.for arcsec-level precision also.
- *
- * REFERENCES:
- *  <ol>
- *   <li>Kaplan, G. H. et. al. (1989). Astron. Journ. 97, 1197-1210.</li>
- *   <li>Kaplan, G. H. (2003), 'Another Look at Non-Rotating Origins', Proceedings of IAU
- *   XXV Joint Discussion 16.</li>
- *  </ol>
- *
- * @param jd_tt_high    [day] High-order part of Terrestrial Time (TT) based Julian date.
- * @param jd_tt_low     [day] Low-order part of Terrestrial Time (TT) based Julian date.
- * @param ut1_to_tt     [s] TT - UT1 Time difference in seconds
- * @param accuracy      NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param xp            [arcsec] Conventionally-defined X coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param yp            [arcsec] Conventionally-defined Y coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param in            Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to ITRS axes (terrestrial system)
- * @param[out] out      Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to True of Date (TOD) axes (celestial system)
- * @return              0 if successful, -1 if either of the vector arguments is NULL, 1 if
- *                      'accuracy' is invalid, or else 10 + the error from cio_location(), or
- *                      20 + error from cio_basis().
- *
- * @sa itrs_to_cirs()
- * @sa tod_to_itrs()
- * @sa tod_to_j2000()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int itrs_to_tod(double jd_tt_high, double jd_tt_low, double ut1_to_tt, enum novas_accuracy accuracy, double xp, double yp, const double *in,
-        double *out) {
-  prop_error("itrs_to_tod",
-          ter2cel(jd_tt_high, jd_tt_low - ut1_to_tt / DAY, ut1_to_tt, EROT_GST, accuracy, NOVAS_DYNAMICAL_CLASS, xp, yp, in, out), 0);
-  return 0;
-}
-
-/**
- * Rotates a position vector from the dynamical CIRS frame of date to the Earth-fixed ITRS frame
- * (IAU 2000 standard method).
- *
- * If both 'xp' and 'yp' are set to 0 no polar motion is included in the transformation.
- *
- * If extreme (sub-microarcsecond) accuracy is not required, you can use UT1-based Julian date
- * instead of the TT-based Julian date and set the 'ut1_to_tt' argument to 0.0. and you can
- * use UTC-based Julian date the same way.for arcsec-level precision also.
- *
- *
- * REFERENCES:
- *  <ol>
- *   <li>Kaplan, G. H. et. al. (1989). Astron. Journ. 97, 1197-1210.</li>
- *   <li>Kaplan, G. H. (2003), 'Another Look at Non-Rotating Origins', Proceedings of IAU XXV
- *   Joint Discussion 16.</li>
- *  </ol>
- *
- * @param jd_tt_high    [day] High-order part of Terrestrial Time (TT) based Julian date.
- * @param jd_tt_low     [day] Low-order part of Terrestrial Time (TT) based Julian date.
- * @param ut1_to_tt     [s] TT - UT1 Time difference in seconds
- * @param accuracy      NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param xp            [arcsec] Conventionally-defined X coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param yp            [arcsec] Conventionally-defined Y coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param in            Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to CIRS axes (celestial system).
- * @param[out] out      Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to ITRS axes (terrestrial system).
- * @return              0 if successful, -1 if either of the vector arguments is NULL, 1 if
- *                      'accuracy' is invalid, 2 if 'method' is invalid 10--20, 3 if the method
- *                      and option are mutually incompatible, or else 10 + the error from
- *                      cio_location(), or 20 + error from cio_basis().
- *
- * @sa tod_to_itrs()
- * @sa itrs_to_cirs()
- * @sa gcrs_to_cirs()
- * @sa cirs_to_gcrs()
- * @sa cirs_to_tod()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int cirs_to_itrs(double jd_tt_high, double jd_tt_low, double ut1_to_tt, enum novas_accuracy accuracy, double xp, double yp,
-        const double *in, double *out) {
-  prop_error("cirs_to_itrs",
-          cel2ter(jd_tt_high, jd_tt_low - ut1_to_tt / DAY, ut1_to_tt, EROT_ERA, accuracy, NOVAS_DYNAMICAL_CLASS, xp, yp, in, out), 0);
-  return 0;
-}
-
-/**
- * Rotates a position vector from the dynamical True of Date (TOD) frame of date the Earth-fixed
- * ITRS frame (pre IAU 2000 method).
- *
- * If both 'xp' and 'yp' are set to 0 no polar motion is included in the transformation.
- *
- * If extreme (sub-microarcsecond) accuracy is not required, you can use UT1-based Julian date
- * instead of the TT-based Julian date and set the 'ut1_to_tt' argument to 0.0. and you can
- * use UTC-based Julian date the same way.for arcsec-level precision also.
- *
- * REFERENCES:
- *  <ol>
- *   <li>Kaplan, G. H. et. al. (1989). Astron. Journ. 97, 1197-1210.</li>
- *   <li>Kaplan, G. H. (2003), 'Another Look at Non-Rotating Origins', Proceedings of IAU XXV
- *   Joint Discussion 16.</li>
- *  </ol>
- *
- * @param jd_tt_high    [day] High-order part of Terrestrial Time (TT) based Julian date.
- * @param jd_tt_low     [day] Low-order part of Terrestrial Time (TT) based Julian date.
- * @param ut1_to_tt     [s] TT - UT1 Time difference in seconds.
- * @param accuracy      NOVAS_FULL_ACCURACY (0) or NOVAS_REDUCED_ACCURACY (1)
- * @param xp            [arcsec] Conventionally-defined X coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param yp            [arcsec] Conventionally-defined Y coordinate of celestial intermediate
- *                      pole with respect to ITRS pole, in arcseconds.
- * @param in            Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to True of Date (TOD) axes (celestial system).
- * @param[out] out      Position vector, geocentric equatorial rectangular coordinates,
- *                      referred to ITRS axes (terrestrial system).
- * @return              0 if successful, -1 if either of the vector arguments is NULL, 1 if
- *                      'accuracy' is invalid, 2 if 'method' is invalid 10--20, 3 if the method
- *                      and option are mutually incompatible, or else 10 + the error from
- *                      cio_location(), or 20 + error from cio_basis().
- *
- * @sa cirs_to_itrs()
- * @sa itrs_to_tod()
- * @sa j2000_to_tod()
- * @sa tod_to_gcrs()
- * @sa tod_to_j2000()
- * @sa tod_to_cirs()
- *
- * @since 1.0
- * @author Attila Kovacs
- */
-int tod_to_itrs(double jd_tt_high, double jd_tt_low, double ut1_to_tt, enum novas_accuracy accuracy, double xp, double yp, const double *in,
-        double *out) {
-  prop_error("tod_to_itrs",
-          cel2ter(jd_tt_high, jd_tt_low - ut1_to_tt / DAY, ut1_to_tt, EROT_GST, accuracy, NOVAS_DYNAMICAL_CLASS, xp, yp, in, out), 0);
-  return 0;
-}
 
 /**
  * Computes the gravitationally undeflected position of an observed source position due to the
@@ -1111,6 +282,7 @@ int make_ephem_object(const char *name, long num, object *body) {
   return 0;
 }
 
+
 /**
  * Sets a celestial object to be a Solar-system orbital body. Typically this would be used to define
  * minor planets, asteroids, comets, or even planetary satellites.
@@ -1139,6 +311,48 @@ int make_orbital_object(const char *name, long num, const novas_orbital *orbit, 
     return novas_error(-1, EINVAL, fn, "Input orbital elements is NULL");
   prop_error(fn, (make_object(NOVAS_ORBITAL_OBJECT, num, name, NULL, body) ? -1 : 0), 0);
   body->orbit = *orbit;
+  return 0;
+}
+
+/**
+ * Sets the orientation of an orbital system using the RA and DEC coordinates of the pole
+ * of the Laplace (or else equatorial) plane relative to which the orbital elements are
+ * defined. Orbital parameters of planetary satellites normally include the R.A. and
+ * declination of the pole of the local Laplace plane in which the Keplerian orbital elements
+ * are referenced.
+ *
+ * The system will become referenced to the equatorial plane, the relative obliquity is set
+ * to (90&deg; - `dec`), while the argument of the ascending node ('Omega') is set to
+ * (90&deg; + `ra`).
+ *
+ * NOTES:
+ * <ol>
+ * <li>You should not expect much precision from the long-range orbital approximations for
+ * planetary satellites. For applications that require precision at any level, you should rely
+ * on appropriate ephemerides, or else on up-to-date short-term orbital elements.</li>
+ * </ol>
+ *
+ * @param type  Coordinate reference system in which `ra` and `dec` are defined (e.g. NOVAS_GCRS).
+ * @param ra    [h] the R.A. of the pole of the oribtal reference plane.
+ * @param dec   [deg] the declination of the pole of the oribtal reference plane.
+ * @param[out]  sys   Orbital system
+ * @return      0 if successful, or else -1 (errno will be set to EINVAL) if the output `sys`
+ *              pointer is NULL.
+ *
+ * @author Attila Kovacs
+ * @since 1.2
+ *
+ * @sa make_orbital_object()
+ */
+int novas_set_orbsys_pole(enum novas_reference_system type, double ra, double dec, novas_orbital_system *restrict sys) {
+  if (!sys)
+    return novas_error(-1, EINVAL, "novas_set_orbsys_pole", "input system is NULL");
+
+  sys->plane = NOVAS_EQUATORIAL_PLANE;
+  sys->type = type;
+  sys->obl = remainder(90.0 - dec, DEG360);
+  sys->Omega = remainder(15.0 * ra + 90.0, DEG360);
+
   return 0;
 }
 
@@ -1319,157 +533,6 @@ int make_solar_system_observer(const double *sc_pos, const double *sc_vel, obser
 }
 
 /**
- * Converts a radial recession velocity to a redshift value (z = &delta;f / f<sub>rest</sub>).
- * It is based on the relativistic formula:
- * <pre>
- *  1 + z = sqrt((1 + &beta;) / (1 - &beta;))
- * </pre>
- * where &beta; = v / c.
- *
- * @param vel   [km/s] velocity (i.e. rate) of recession.
- * @return      the corresponding redshift value (&delta;&lambda; / &lambda;<sub>rest</sub>), or NAN if
- *              the input velocity is invalid (i.e., it exceeds the speed of light).
- *
- * @sa novas_z2v()
- * @sa novas_z_add()
- *
- * @author Attila Kovacs
- * @since 1.2
- */
-double novas_v2z(double vel) {
-  vel *= NOVAS_KMS / C;   // [km/s] -> beta
-  if(fabs(vel) > 1.0) {
-    novas_error(-1, EINVAL, "novas_v2z", "velocity exceeds speed of light v=%g km/s", vel * C / NOVAS_KMS);
-    return NAN;
-  }
-  return sqrt((1.0 + vel) / (1.0 - vel)) - 1.0;
-}
-
-/**
- * Returns the gravitational redshift (_z_) for light emitted near a massive spherical body
- * at some distance from its center, and observed at some very large (infinite) distance away.
- *
- * @param M_kg    [kg] Mass of gravitating body that is contained inside the emitting radius.
- * @param r_m     [m] Radius at which light is emitted.
- * @return        The gravitational redshift (_z_) for an observer at very large  (infinite)
- *                distance from the gravitating body.
- *
- * @sa redshift_vrad()
- * @sa unredshift_vrad()
- * @sa novas_z_add()
- *
- * @since 1.2
- * @author Attila Kovacs
- */
-double grav_redshift(double M_kg, double r_m) {
-  static const double twoGoverC2 = 2.0 * 6.6743e-11 / (C * C); // 2G/c^2 in SI units.
-  return 1.0 / sqrt(1.0 - twoGoverC2 * M_kg / r_m) - 1.0;
-}
-
-/**
- * Applies an incremental redshift correction to a radial velocity. For example, you may
- * use this function to correct a radial velocity calculated by `rad_vel()` or `rad_vel2()`
- * for a Solar-system body to account for the gravitational redshift for light originating
- * at a specific distance away from the body. For the Sun, you may want to undo the redshift
- * correction applied for the photosphere using `unredshift_vrad()` first.
- *
- * @param vrad    [km/s] Radial velocity
- * @param z       Redshift correction to apply
- * @return        [km/s] The redshift corrected radial velocity or NAN if the redshift value
- *                is invalid (errno will be set to EINVAL).
- *
- * @sa unredshift_vrad()
- * @sa grav_redshift()
- * @sa novas_z_add()
- *
- * @since 1.2
- * @author Attila Kovacs
- */
-double redshift_vrad(double vrad, double z) {
-  static const char *fn = "redshift_vrad";
-  double z0;
-  if(z <= -1.0) {
-    novas_error(-1, EINVAL, fn, "invalid redshift value: z=%g", z);
-    return NAN;
-  }
-  z0 = novas_v2z(vrad);
-  if(isnan(z0)) novas_trace(fn, -1, 0);
-  return novas_z2v((1.0 + z0) * (1.0 + z) - 1.0);
-}
-
-/**
- * Undoes an incremental redshift correction that was applied to radial velocity.
- *
- * @param vrad    [km/s] Radial velocity
- * @param z       Redshift correction to apply
- * @return        [km/s] The radial velocity without the redshift correction or NAN if the
- *                redshift value is invalid. (errno will be set to EINVAL)
- *
- * @sa redshift_vrad()
- * @sa grav_redshift()
- *
- * @since 1.2
- * @author Attila Kovacs
- */
-double unredshift_vrad(double vrad, double z) {
-  static const char *fn = "unredshift_vrad";
-  double z0;
-  if(z <= -1.0) {
-    novas_error(-1, EINVAL, fn, "invalid redshift value: z=%g", z);
-    return NAN;
-  }
-  z0 = novas_v2z(vrad);
-  if(isnan(z0)) novas_trace(fn, -1, 0);
-  return novas_z2v((1.0 + z0) / (1.0 + z) - 1.0);
-}
-
-/**
- * Compounds two redshift corrections, e.g. to apply (or undo) a series gravitational redshift
- * corrections and/or corrections for a moving observer. It's effectively using
- * (1 + z) = (1 + z1) * (1 + z2).
- *
- * @param z1    One of the redshift values
- * @param z2    The other redshift value
- * @return      The compound redshift value, ot NAN if either input redshift is invalid (errno
- *              will be set to EINVAL).
- *
- * @sa grav_redshift()
- * @sa redshift_vrad()
- * @sa unredshift_vrad()
- *
- * @since 1.2
- * @author Attila Kovacs
- */
-double novas_z_add(double z1, double z2) {
-  if(z1 <= -1.0 || z2 <= -1.0) {
-    novas_error(-1, EINVAL, "novas_z_add", "invalid redshift value: z1=%g, z2=%g", z1, z2);
-    return NAN;
-  }
-  return z1 + z2 + z1 * z2;
-}
-
-/**
- * Returns the inverse of a redshift value, that is the redshift for a body moving with the same
- * velocity as the original but in the opposite direction.
- *
- * @param z     A redhift value
- * @return      The redshift value for a body moving in the opposite direction with the
- *              same speed, or NAN if the input redshift is invalid.
- *
- * @sa novas_z_add()
- *
- * @since 1.2
- * @author Attila Kovacs
- */
-double novas_z_inv(double z) {
-  if(z <= -1.0) {
-    novas_error(-1, EINVAL, "novas_z_inv", "invalid redshift value: z=%g", z);
-    return NAN;
-  }
-  return 1.0 / (1.0 + z) - 1.0;
-}
-
-/**
  * Returns the NOVAS planet ID for a given name (case insensitive), or -1 if no match is found.
  *
  * @param name    The planet name, or that for the "Sun", "Moon" or "SSB" (case insensitive).
@@ -1514,47 +577,7 @@ enum novas_planet novas_planet_for_name(const char *restrict name) {
   return novas_error(-1, EINVAL, fn, "No match for name: '%s'", name);
 }
 
-/**
- * Sets the orientation of an orbital system using the RA and DEC coordinates of the pole
- * of the Laplace (or else equatorial) plane relative to which the orbital elements are
- * defined. Orbital parameters of planetary satellites normally include the R.A. and
- * declination of the pole of the local Laplace plane in which the Keplerian orbital elements
- * are referenced.
- *
- * The system will become referenced to the equatorial plane, the relative obliquity is set
- * to (90&deg; - `dec`), while the argument of the ascending node ('Omega') is set to
- * (90&deg; + `ra`).
- *
- * NOTES:
- * <ol>
- * <li>You should not expect much precision from the long-range orbital approximations for
- * planetary satellites. For applications that require precision at any level, you should rely
- * on appropriate ephemerides, or else on up-to-date short-term orbital elements.</li>
- * </ol>
- *
- * @param type  Coordinate reference system in which `ra` and `dec` are defined (e.g. NOVAS_GCRS).
- * @param ra    [h] the R.A. of the pole of the oribtal reference plane.
- * @param dec   [deg] the declination of the pole of the oribtal reference plane.
- * @param[out]  sys   Orbital system
- * @return      0 if successful, or else -1 (errno will be set to EINVAL) if the output `sys`
- *              pointer is NULL.
- *
- * @author Attila Kovacs
- * @since 1.2
- *
- * @sa make_orbital_object()
- */
-int novas_set_orbsys_pole(enum novas_reference_system type, double ra, double dec, novas_orbital_system *restrict sys) {
-  if (!sys)
-    return novas_error(-1, EINVAL, "novas_set_orbsys_pole", "input system is NULL");
 
-  sys->plane = NOVAS_EQUATORIAL_PLANE;
-  sys->type = type;
-  sys->obl = remainder(90.0 - dec, DEG360);
-  sys->Omega = remainder(15.0 * ra + 90.0, DEG360);
-
-  return 0;
-}
 
 /**
  * Returns the horizontal Parallactic Angle (PA) calculated for a gorizontal Az/El location of the sky. The PA
@@ -1844,176 +867,3 @@ int novas_xyz_to_uvw(const double *xyz, double ha, double dec, double *uvw) {
   return 0;
 }
 
-static int convert_lsr_ssb_vel(const double *vLSR, int sign, double *vSSB) {
-  static const double betaSSB[3] = { 11.1 * NOVAS_KMS / C, 12.24 * NOVAS_KMS / C, 7.25 * NOVAS_KMS / C };
-  int i;
-
-  for(i = 3; --i >= 0; )
-    vSSB[i] = novas_add_beta(vLSR[i] * NOVAS_KMS / C, sign * betaSSB[i]) * C / NOVAS_KMS;
-
-  return 0;
-}
-
-/**
- * Returns a Solar System Baricentric (SSB) radial velocity for a radial velocity that is referenced to the
- * Local Standard of Rest (LSR). Internally, NOVAS always uses barycentric radial velocities, but it
- * is just as common to have catalogs define radial velocities referenced to the LSR.
- *
- * The SSB motion w.r.t. the barycenter is assumed to be (11.1, 12.24, 7.25) km/s in ICRS (Shoenrich et al.
- * 2010).
- *
- * REFERENCES:
- * <ol>
- * <li>Ralph Schoenrich, James Binney, Walter Dehnen, Monthly Notices of the Royal Astronomical Society,
- * Volume 403, Issue 4, April 2010, Pages 1829–1833, https://doi.org/10.1111/j.1365-2966.2010.16253.x</li>
- * </ol>
- *
- * @param epoch     [yr] Coordinate epoch in which the coordinates and velocities are defined. E.g. 2000.0.
- * @param ra        [h] Right-ascenscion of source at given epoch.
- * @param dec       [deg] Declination of source at given epoch.
- * @param vLSR      [km/s] radial velocity defined against the Local Standard of Rest (LSR), at given epoch.
- *
- * @return          [km/s] Equivalent Solar-System Barycentric radial velocity.
- *
- * @since 1.3
- * @author Attila Kovacs
- *
- * @sa make_cat_entry()
- * @sa novas_ssb_to_lsr_vel()
- */
-double novas_lsr_to_ssb_vel(double epoch, double ra, double dec, double vLSR) {
-  double u[3] = {}, v[3];
-  double jd = NOVAS_JD_J2000 + 365.25 * (epoch - 2000.0);
-  int i;
-
-  radec2vector(ra, dec, 1.0, u);
-  for(i = 3; --i >= 0; )
-    v[i] = vLSR * u[i];
-
-  precession(jd, v, NOVAS_JD_J2000, v);
-  convert_lsr_ssb_vel(v, -1, v);
-  precession(NOVAS_JD_J2000, v, jd, v);
-
-  return novas_vdot(u, v);
-}
-
-/**
- * Returns a radial-velocity referenced to the Local Standard of Rest (LSR) for a given Solar-System
- * Barycentric (SSB) radial velocity. Internally, NOVAS always uses barycentric radial velocities, but it
- * is just as common to have catalogs define radial velocities referenced to the LSR.
- *
- * The SSB motion w.r.t. the barycenter is assumed to be (11.1, 12.24, 7.25) km/s in ICRS (Shoenrich et al.
- * 2010).
- *
- * REFERENCES:
- * <ol>
- * <li>Ralph Schoenrich, James Binney, Walter Dehnen, Monthly Notices of the Royal Astronomical Society,
- * Volume 403, Issue 4, April 2010, Pages 1829–1833, https://doi.org/10.1111/j.1365-2966.2010.16253.x</li>
- * </ol>
- *
- * @param epoch     [yr] Coordinate epoch in which the coordinates and velocities are defined. E.g. 2000.0.
- * @param ra        [h] Right-ascenscion of source at given epoch.
- * @param dec       [deg] Declination of source at given epoch.
- * @param vLSR      [km/s] radial velocity defined against the Local Standard of Rest (LSR), at given epoch.
- *
- * @return          [km/s] Equivalent Solar-System Barycentric radial velocity.
- *
- * @since 1.3
- * @author Attila Kovacs
- *
- * @sa make_cat_entry()
- * @sa novas_lsr_to_ssb_vel()
- */
-double novas_ssb_to_lsr_vel(double epoch, double ra, double dec, double vLSR) {
-  double u[3] = {}, v[3];
-  double jd = NOVAS_JD_J2000 + 365.25 * (epoch - 2000.0);
-  int i;
-
-  radec2vector(ra, dec, 1.0, u);
-  for(i = 3; --i >= 0;)
-    v[i] = vLSR * u[i];
-
-  precession(jd, v, NOVAS_JD_J2000, v);
-  convert_lsr_ssb_vel(v, 1, v);
-  precession(NOVAS_JD_J2000, v, jd, v);
-
-  return novas_vdot(u, v);
-}
-
-/**
- * Returns the Julian day corresponding to an astronomical coordinate epoch.
- *
- * @param system        Coordinate system, e.g. "ICRS", "B1950.0", "J2000.0", "FK4", "FK5",
- *                      "1950", "2000", or "HIP". In general, any Besselian or Julian year epoch
- *                      can be used by year (e.g. "B1933.193" or "J2022.033"), or else the fixed
- *                      values listed. If 'B' or 'J' is ommitted in front of the epoch year, then
- *                      Besselian epochs are assumed prior to 1984.0.
- * @return              [day] The Julian day corresponding to the given coordinate epoch, or else
- *                      NAN if the input string is NULL or the input is not recognised as a
- *                      coordinate epoch specification (errno will be set to EINVAL).
- *
- * @since 1.3
- * @author Attila Kovacs
- *
- * @sa make_cat_object_sys()
- * @sa make_redshifted_object_sys()
- * @sa transform_cat()
- * @sa precession()
- * @sa NOVAS_SYSTEM_ICRS
- * @sa NOVAS_SYSTEM_B1950
- * @sa NOVAS_SYSTEM_J2000
- * @sa NOVAS_SYSTEM_HIP
- */
-double novas_epoch(const char *restrict system) {
-  static const char *fn = "novas_epoch";
-
-  double year;
-  char type = 0, *tail = NULL;
-
-  if(!system) {
-    novas_error(0, EINVAL, fn, "epoch is NULL");
-    return NAN;
-  }
-
-  if(!system[0]) {
-    novas_error(0, EINVAL, fn, "epoch is empty");
-    return NAN;
-  }
-
-  if(strcasecmp(system, NOVAS_SYSTEM_ICRS) == 0)
-    return NOVAS_JD_J2000;
-
-  if(strcasecmp(system, NOVAS_SYSTEM_FK5) == 0)
-    return NOVAS_JD_J2000;
-
-  if(strcasecmp(system, NOVAS_SYSTEM_FK4) == 0)
-    return NOVAS_JD_B1950;
-
-  if(strcasecmp(system, NOVAS_SYSTEM_HIP) == 0)
-    return NOVAS_JD_HIP;
-
-  if(toupper(system[0]) == 'B') {
-    type = 'B';
-    system++;
-  }
-  else if(toupper(system[0]) == 'J') {
-    type = 'J';
-    system++;
-  }
-
-  errno = 0;
-
-  year = strtod(system, &tail);
-  if(tail <= system) {
-    novas_error(0, EINVAL, fn, "Invalid epoch: %s", system);
-    return NAN;
-  }
-
-  if(!type)
-    type = year < 1984.0 ? 'B' : 'J';
-
-  if(type == 'J')
-    return NOVAS_JD_J2000 + (year - 2000.0) * JULIAN_YEAR_DAYS;
-
-  return NOVAS_JD_B1950 + (year - 1950.0) * BESSELIAN_YEAR_DAYS;
-}
