+++ test/Makefile
@@ -19,11 +19,13 @@ CHECKOPTS = --enable=performance,warning,portability --language=c \
 
 #OBJECTS := $(subst obj/,,$(OBJECTS))
 OBJECTS := novas.o nutation.o solsys3.o super.o frames.o timescale.o \
-           refract.o naif.o parse.o
+           transform.o place.o spectral.o refract.o naif.o parse.o
+
 TESTS := test-compat test-cio_file test-super test-errors
+
 COVFILES := novas.c.gcov nutation.c.gcov solsys3.c.gcov super.c.gcov \
-            timescale.c.gcov refract.c.gcov frames.c.gcov naif.c.gcov \
-            parse.c.gcov
+            transform.c.gcov place.c.gcov timescale.c.gcov spectral.c.gcov \
+            refract.c.gcov frames.c.gcov naif.c.gcov parse.c.gcov
 
 ifeq ($(CALCEPH_SUPPORT), 1)
   LDFLAGS += -lcalceph
